many changes.
added preview for tweet text changes.
used new SeekBarPreference.
profile image, color label and user name in account selector.
removed features not avaliable in twitter API version 1.1.
support RTL color label in AR locale.
used since_id for tweet loading, this may faster and uses less network traffic.
added "profile banner image", looks very nice :)

+++ src/org/mariotaku/twidere/activity/DraftsActivity.java
@@ -82,7 +82,7 @@ public class DraftsActivity extends BaseActivity implements LoaderCallbacks<Curs
 		mResolver = getContentResolver();
 		mService = getTwidereApplication().getServiceInterface();
 		mPreferences = getSharedPreferences(SHARED_PREFERENCES_NAME, Context.MODE_PRIVATE);
-		mTextSize = mPreferences.getFloat(PREFERENCE_KEY_TEXT_SIZE, PREFERENCE_DEFAULT_TEXT_SIZE);
+		mTextSize = mPreferences.getInt(PREFERENCE_KEY_TEXT_SIZE, PREFERENCE_DEFAULT_TEXT_SIZE);
 		setContentView(R.layout.base_list);
 		getSupportActionBar().setDisplayHomeAsUpEnabled(true);
 		mAdapter = new DraftsAdapter(this);
@@ -183,7 +183,7 @@ public class DraftsActivity extends BaseActivity implements LoaderCallbacks<Curs
 	@Override
 	public void onResume() {
 		super.onResume();
-		final float text_size = mPreferences.getFloat(PREFERENCE_KEY_TEXT_SIZE, PREFERENCE_DEFAULT_TEXT_SIZE);
+		final float text_size = mPreferences.getInt(PREFERENCE_KEY_TEXT_SIZE, PREFERENCE_DEFAULT_TEXT_SIZE);
 		mAdapter.setTextSize(text_size);
 		if (mTextSize != text_size) {
 			mTextSize = text_size;
